backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pandas
joblib
python-multipart
google-generativeai